+++ opensquirrel/passes/router/__init__.py
@@ -1,7 +1,3 @@
 from opensquirrel.passes.router.general_router import Router
-from opensquirrel.passes.router.routing_checker import RoutingChecker
 
-__all__ = [
-    "Router",
-    "RoutingChecker",
-]
+__all__ = ["Router"]