+++ opensquirrel/passes/validator/__init__.py
@@ -2,5 +2,6 @@
 
 from opensquirrel.passes.validator.general_validator import Validator
 from opensquirrel.passes.validator.native_gate_validator import NativeGateValidator
+from opensquirrel.passes.validator.routing_validator import RoutingValidator
 
-__all__ = ["NativeGateValidator", "Validator"]
+__all__ = ["NativeGateValidator", "RoutingValidator", "Validator"]