+++ CHANGELOG.md
@@ -16,6 +16,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - Refactor: removed generators
 
+- Changed the `RoutingChecker` pass to a `RoutingValidator` pass
+
 
 ## [0.3.1] - [ 2025-01-31 ]